openapi package renamed openapi3
