+++ cmd/print.go
@@ -32,7 +32,7 @@ import (
 	_ "github.com/kubernetes-sigs/ingress2gateway/pkg/i2gw/providers/ingressnginx"
 	_ "github.com/kubernetes-sigs/ingress2gateway/pkg/i2gw/providers/istio"
 	_ "github.com/kubernetes-sigs/ingress2gateway/pkg/i2gw/providers/kong"
-	_ "github.com/kubernetes-sigs/ingress2gateway/pkg/i2gw/providers/openapi"
+	_ "github.com/kubernetes-sigs/ingress2gateway/pkg/i2gw/providers/openapi3"
 )
 
 type PrintRunner struct {
