+++ pkg/i2gw/providers/openapi3/fixtures/input/1-petstore3.yaml
