+++ pkg/i2gw/providers/openapi3/fixtures/input/2-hostnames.yaml
