+++ pkg/i2gw/providers/openapi3/fixtures/input/3-parameters.yaml
