+++ pkg/i2gw/providers/openapi3/fixtures/input/4-too-many-rules.json
