+++ pkg/i2gw/providers/openapi3/fixtures/output/1-petstore3.yaml
