+++ pkg/i2gw/providers/openapi3/fixtures/output/2-hostnames.yaml
