+++ pkg/i2gw/providers/openapi3/fixtures/output/3-parameters.yaml
