+++ pkg/i2gw/providers/openapi3/fixtures/output/4-too-many-rules.json
