+++ pkg/i2gw/providers/openapi3/resource_reader.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package openapi
+package openapi3
 
 import (
 	"context"
